Add Ajax Handler class
